requirements.txt
quart
python-dotenv
aiohttp
bs4